Remove unnecessary elif
Early exit means we can avoid the else side of elif, since we've already
exited by then.
This means less indentations overall, repeating ifs are nicely aligned.

I was also told recently: "having an ELSE should be a red flag", and I
kind of agree with that, in this case we didn't need it, can clean up.